WIIS-7727: finished the responsive + added window-size-service + account creation page final update
Convert all the % usage in scss file into rem. All menu now have the expected behaviour for all screen size. Added a service to manage when the size of the application window change. Fix the last bug in the account creation page and change some of the scss to better match the design.