WIIS-7720: modal first version + added feature to component text area and password + fix bug in all parameter page + linked modal to global settings page
Created the form-modal component and made it according to the design. Added a variation of text are and password component where the field name label is on top of the field. ix a small spelling mistake in all parameters page which made it impossible for min and max length error message to appear. Linked the created modal to the global parameters menu to enable the admin username and password edit.