WIIS-7718: migrated plugin + ftp service + file service + small code refacto + REMOVE ALL STRICT MODE POLICY
migrated all ionic-native plugins to their awesome-cordova counterpart to avoid depreciation. Created the ftp service for connecting and uploading file to a ftp server. Created the file service to create and write into file with the desired format. Change the StrictMode policy in the MainActivity.java in order for the ftp plugin to work. This should be changed in the long run (either by recreating the plugin or because the plugin will get updated)