WIIS-7713: Clear Todo + Renamed and Refactor Window Size Service + Deleted Screen orientation service + Active mode page V1
Change a lot of file in order to clear some Todo (mostly linked to connecting page to the active mode page which was not created yet). Refactor Window Size Service and Screen Orientation Service into one and only service named Window Service. Made the first version of the active mode page (not responsive yet and does not respect the design in landscape mode)