WIIS-7718 + WIIS-7722: modal size correction + change the naming of exported file + fix some error in order to build an apk
Change global.scss to better handle the size of the modal depending on screen size and orientation. File exported by the sftp service now have the id of the device in their name in order to avoid same with the same name on the receiving server. Fix some errors in a lot of component/page (mostly typing problems) that where blocking the creation of the first apk.