WIIS-7719: modal created and functional + cleaned some TODO + cleaned import + minor bug fix + delay between two clocking is taking into account
Created the clocking-info-modal page and the clocking-info-line component. Modal are now open when a badge is clocked. Cleaned some Todo by updating the code of some file. Remove unused import from a lot of files. Fix the bug where multiple copy of the same modal could be open on top of each other. Application is not reacting to the clocking of a badge if it's last clocking is younger than the delay between two clocking parameter.