WIIS-8125 + WIIS-8149: Application icon and splash screen + Finished the design fix in the active mode page + Application name
Added a splash screen that match the application icon and used it when app is started. Renamed the app to WiiTime. Had to change the scss file of the active mode page again to match the requested design. Changed the scss file of the clocking info line component to be more responsive.