WIIS-7732: Added toast service + added the toast in all the place in which they were required
Created the toast service file and it's content. Added the toast for multiple clocking of the same badge, the toast for the export clocking data result, the toast to confirm the save of the change to the param and the toast which display the result of the SFTP server connexion test.