WIIS-7746: Fix audio service + correct version number displayed + refacto home page + added warning in export + removed screen waking service
Change the audio service to better handle the load and unload of sound and fix a bug related to that. Change the home page design to look like other application page (also translated its text to french). Added a warning toast when exporting more than 10000 clocking at once with the button in the clocking settings component. Screen is able to go asleep again as it was consuming too much energy to keep it awake all the time.